requirements.txt
bs4
deep-translator
g4f
gradio_client
pillow
PySocks
psutil